♻️ 8F6BQR task: decompose policy taxonomy types
