+++ .agentplane/tasks/202605290636-8F6BQR/pr/diffstat.txt
@@ -1,0 +1,3 @@
+ packages/agentplane/src/policy/taxonomy-types.ts | 35 ++++++++++++++++++++++
+ packages/agentplane/src/policy/taxonomy.ts       | 38 ++----------------------
+ 2 files changed, 38 insertions(+), 35 deletions(-)
