+++ .agentplane/tasks/202605290636-8F6BQR/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 8F6BQR task: Policy taxonomy type decomposition [202605290636-8F6BQR]
