+++ .agentplane/tasks/202605290636-8F6BQR/pr/review.md
@@ -1,0 +1,38 @@
+# PR Review
+
+Created: 2026-05-29T06:36:36.542Z
+
+## Task
+
+- Task: `202605290636-8F6BQR`
+- Title: Policy taxonomy type decomposition
+- Status: DOING
+- Branch: `task/202605290636-8F6BQR/policy-taxonomy-type-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290636-8F6BQR/README.md`
+
+## Verification
+
+- State: ok
+- Note: KnownPolicyActionId moved to packages/agentplane/src/policy/taxonomy-types.ts and re-exported from taxonomy.ts; descriptor data and resolvePolicyActionDescriptor behavior are unchanged. Checks passed: bun test packages/agentplane/src/policy/taxonomy.test.ts (4 pass); bun run typecheck; bun run arch:check; bun run knip:check; bun run lint:core; bun run format:changed; bun run hotspots:check (runtime hotspots 2 -> 1).
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T06:36:36.542Z
+- Branch: task/202605290636-8F6BQR/policy-taxonomy-type-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ packages/agentplane/src/policy/taxonomy-types.ts | 35 ++++++++++++++++++++++
+ packages/agentplane/src/policy/taxonomy.ts       | 38 ++----------------------
+ 2 files changed, 38 insertions(+), 35 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
